[DMS-1124] Extract shared aligned-base-scope resolution helper
Deduplicate the aligned-base-scope path resolution logic that was
identically implemented in CompiledScopeAdapterFactory and
RelationalWriteMerge into a shared JsonScopePathHelper utility
in Backend.External.
